Исправлена ошибка с uint8 и float

+++ lab_2_np/np_episode_1/np_episode_2.py
@@ -5,8 +5,8 @@ for j in range(3):
     img = Image.open(f'lunar_images/lunar0{j+1}_raw.jpg')
     data = np.array(img)
 
-    a = data.min()
-    b = data.max()
+    a = float(data.min())
+    b = float(data.max())
     
     k = 255 / (b - a)
     p = (255 - k * (a + b)) / 2
@@ -14,7 +14,7 @@ for j in range(3):
     rows = data.shape[0]
     columns = data.shape[1]
     
-    updated_data = np.empty_like(data)
+    updated_data = np.empty_like(data, dtype = np.uint8)
     
     # print(rows)
     # print(columns)
